+++ src/main/java/com/Sucat/domain/board/controller/BoardController.java
@@ -1,43 +1,40 @@
 package com.Sucat.domain.board.controller;
 
-import com.Sucat.domain.board.DTO.BoardPostRequestDTO;
-import com.Sucat.domain.board.DTO.BoardResponse;
-import com.Sucat.domain.board.DTO.BoardUpdateRequestDTO;
-import com.Sucat.domain.board.DTO.ResponseDTO;
+import com.Sucat.domain.board.dto.BoardPostRequestDTO;
+import com.Sucat.domain.board.dto.BoardResponse;
+import com.Sucat.domain.board.dto.BoardUpdateRequestDTO;
+import com.Sucat.domain.board.dto.ResponseDTO;
 import com.Sucat.domain.board.model.BoardCategory;
 import com.Sucat.domain.board.service.BoardService;
 import com.Sucat.global.common.code.SuccessCode;
 import com.Sucat.global.common.response.ApiResponse;
 import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Objects;
 
 @RestController
-@RequestMapping("/api/boards")
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/boards")
 public class BoardController {
 
     private final BoardService boardService;
 
-    @Autowired
-    public BoardController(BoardService boardService) {
-        this.boardService = boardService;
-    }
-
     /* 게시글 작성 */
     @PostMapping
-    public ResponseEntity<ApiResponse<Objects>> createBoard(@RequestBody BoardPostRequestDTO boardPostRequestDTO,
-                                                            HttpServletRequest request) {
+    public ResponseEntity<ApiResponse<Objects>> createBoard(@RequestBody BoardPostRequestDTO boardPostRequestDTO, HttpServletRequest request) {
         boardService.createBoard(boardPostRequestDTO, request);
-        return ApiResponse.onSuccess(SuccessCode._OK);
+        return ApiResponse.onSuccess(SuccessCode._CREATED);
     }
 
     /* (자유, 비밀, 중고장터) 카테고리의 게시글 목록 조회 */
     @GetMapping
-    public ResponseDTO getBoardsByCategory(@RequestParam BoardCategory category) {
-        return boardService.getAllBoards(category);
+    public ResponseEntity<ApiResponse<Objects>> getBoardsByCategory(@RequestParam BoardCategory category) {
+        ResponseDTO allBoards = boardService.getAllBoards(category);
+
+        return ApiResponse.onSuccess(SuccessCode._OK, allBoards);
     }
 
     /* 게시글 단일 조회 */
