+++ src/main/java/com/Sucat/domain/board/dto/BoardPostDTO.java
@@ -1,4 +1,4 @@
-package com.Sucat.domain.board.DTO;
+package com.Sucat.domain.board.dto;
 
 public class BoardPostDTO {
     private String minute;
