+++ src/main/java/com/Sucat/domain/board/dto/BoardUpdateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.Sucat.domain.board.DTO;
+package com.Sucat.domain.board.dto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
