+++ src/main/java/com/Sucat/domain/board/dto/ResponseDTO.java
@@ -1,4 +1,4 @@
-package com.Sucat.domain.board.DTO;
+package com.Sucat.domain.board.dto;
 
 import java.util.List;
